Split WalkPlan class
